Add Visual Studio 2019 Tools to Windows Image (#1143)
* Add Visual Studio 2019 Tools to Windows Image

Adds Visual Studio 2019 Build Tools to the Windows image. This will
allow building native applications/DLLs for Windows in the Jenkins
infrastructure.

* Make sure install process completes

Makes sure that the install process has completed before moving on.

* Add check for 2019 when installing VS Build Tools

Adds a check for the environment variable (AGENT_OS_VERSION) to only install VS Build Tools if it is 2019.

* Skip goss check on 2022

Skips the goss check fir VS Built Tools if building for 2022.